requirements.txt
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
onnx>=1.14.0
onnxruntime>=1.15.0
scipy>=1.10.0
matplotlib>=3.7.0
pytest>=7.0.0
